# src/engine/server/docs/swagger.yaml
basePath: '{{.BasePath}}'
definitions:
...
  util.CronSchedule:
    properties:
      value:
        type: string
    type: object
  util.JobScheduleResult:
    properties:
      jobSchedules:
        items:
          type: JobSchedule
        type: array
      result:
        $ref: '#/definitions/util.Result'
        type: object
    type: object
...
  util.Result:
    properties:
      code:
        type: integer
      data:
        items:
          type: string
        type: array
      messages:
        items:
          type: Message
        type: array
    type: object
...
host: '{{.Host}}'
info:
  contact:
    email: keith.tenzer@gmail.com
    name: Keith Tenzer
    url: http://www.keithtenzer.com
  description: |-
    APIs for managing Fossul workflows, jobs, profile, and configs
    JSON API definition can be retrieved at <a href="/api/v1/swagger/doc.json">/api/v1/swagger/doc.json</a>
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fossul Framework Server API
  version: "1.0"
paths:
...
  /addSchedule/{profileName}/{configName}/{policy}:
    post:
      consumes:
      - application/json
      description: Add job schedule
      parameters:
      - description: name of profile
        in: path
        name: profileName
        required: true
        type: string
      - description: name of config
        in: path
        name: configName
        required: true
        type: string
      - description: policy name
        in: path
        name: policy
        required: true
        type: string
      - description: 'value: min,hour,dayOfMonth,month,dayOfWeek'
        in: body
        name: cronSchedule
        required: true
        schema:
          $ref: '#/definitions/util.CronSchedule'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Result'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
...
  /deleteSchedule/{profileName}/{configName}/{policy}:
    get:
      consumes:
      - application/json
      description: Delete job schedule
      parameters:
      - description: name of profile
        in: path
        name: profileName
        required: true
        type: string
      - description: name of config
        in: path
        name: configName
        required: true
        type: string
      - description: policy name
        in: path
        name: policy
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.Result'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
...
  /listSchedules:
    get:
      consumes:
      - application/json
      description: List job schedules
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            string:
              description: OK
              type: string
          schema:
            $ref: '#/definitions/util.JobScheduleResult'
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Not Found
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
...
swagger: "2.0"